+++ docs_src/src/guide/changelog.md
@@ -38,6 +38,8 @@ meta:
 - Protect the database instance with mutex.
 - Recovery middleware now correctly handles panics with a `nil` value.
 - The following rules now pass if the validated data type is not supported: `greater_than`, `greater_than_equal`, `lower_than`, `lower_than_equal`, `size`.
+- Type-dependent rules now try to determine what is the expected type by looking up in the rule set for a type rule. If no type rule is present, falls back to the inputted type. This change makes it so the validation message is correct even if the client didn't input the expected type.
+- Fixed a bug triggering a panic if the client inputted a non-array value in an array-validated field.
 
 ## v2.10.x